UnoBot: parts of overlong hand get lost => panic if no visible card matches
Sometimes, the NOTICE sent by the Uno Game Bot with the contents of the
player's hand is too long and gets chopped off by the server. If that
happens and none of the hand cards we *could* understand matches, panic
(play a WD4, play a WILD, draw and pass) and hope for the next round.