UnoBot: adjust skipping next player if next-but-one is winning
If the next-but-one player (bot+2) was about to win, the bot filtered out
all cards that would jump straight to them. Modify this such that if both
the next player (bot+1) and bot+2 are about to win, filter out cards that
would skip bot+1 only if bot+2 has fewer cards than bot+1, i.e. try to
ensure that the game continues with the player who still has more cards in
their hand.