UnoBot: don't ignore color requests if top card already has wanted color
Previously, if the bot is to honor a color request and the top card already
has the correct color, the bot stops honoring the color before choosing a
card, which means it may immediately change the color. Instead, have the
bot honor color requests until immediately *after* it plays the next card.