ConnectionManager: rewrite event dispatch to catch exceptions centrally
C#'s default event handling dispatches events such that if one handler
throws an exception, the remaining handlers are not called. Rewrite
ConnectionManager's event handling such that each handler's exceptions are
caught and logged and processing continues with the next handler. This
should obviate the need for the two-level event handlers that currently
permeate the SharpIrcBot plugins.