LinkInfo: don't misidentify e-mail addresses as authenticated URLs
An e-mail address prefixed by "http://" can be considered an attempt to
provide a username to authenticate with the website hosted at the mail
domain. For example, http://ondra.hosek@example.com/ can be seen as an
attempt to provide the username "ondra.hosek" to http://example.com. Ignore
words with '@' in them when heuristically detecting links to prevent this;
of course, non-heuristic link detection (when the scheme is explicitly
supplied) allows them.