GrammarGen: alternative/optional weighting
Allow specifying weights for alternatives or optionals (as a prefix in
angled brackets). With optionals, the number is a percentage how often
the contents should be produced. With alternatives, the number is a
relative weight; the random number is generated in the range of the sum
of the weights of each alternative.